Make a post request from curly!
The problem with Slacko is probably that channels API is deprecated and
is conversations API instead. Ditch that to use Curly, but also honestly
the error messages SUCK. The problem we were having was that we were
making GET requests when we should've been making POST request but the
error message we got was an "Auth error".

Sonja found out that "-F" is a parameter given to POST requests!
https://curl.haxx.se/docs/manpage.html#-F

Co-authored-by: Sonja Heinze <sonjaleaheinze@gmail.com>