Improve score computation
Before, having had a chat with someone three weeks ago used to weigh the
same as having had a chat with them a year ago. But after a long coffee
chat bot break, you don't remember who you had a chat with before the
break, whereas you care about not repeating chats from after the break.
So the latter should weigh more than the former, which is what this commit
aims for.